requirements.txt
fastapi
uvicorn
gitpython
radon
pydantic
jinja2
pytest
python-multipart
